Added a new Body interface along with implementations.
Previously request bodies were simple strings that were parsed as JSON, but this did not lend well to the absence of bodies. Most HTTP requests can have bodies, but are not required to. Bodies can also be in more or less any format, so having a Body interface makes it easier to implement support different kinds of requests.